+++ 1.21.4/Common/src/main/java/fuzs/spikyspikes/client/SpikySpikesClient.java
@@ -3,12 +3,27 @@ package fuzs.spikyspikes.client;
 import fuzs.puzzleslib.api.client.core.v1.ClientModConstructor;
 import fuzs.puzzleslib.api.client.core.v1.context.BlockEntityRenderersContext;
 import fuzs.puzzleslib.api.client.core.v1.context.ItemModelsContext;
+import fuzs.puzzleslib.api.client.event.v1.model.ModelLoadingEvents;
+import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
 import fuzs.spikyspikes.SpikySpikes;
+import fuzs.spikyspikes.client.handler.SpikeBlockModelHandler;
 import fuzs.spikyspikes.client.renderer.blockentity.SpikeRenderer;
 import fuzs.spikyspikes.client.renderer.special.SpikeSpecialRenderer;
 import fuzs.spikyspikes.init.ModRegistry;
 
 public class SpikySpikesClient implements ClientModConstructor {
+
+    @Override
+    public void onConstructMod() {
+        registerEventHandlers();
+    }
+
+    private static void registerEventHandlers() {
+        // TODO remove this check once the custom block model is supported for NeoForge
+        if (ModLoaderEnvironment.INSTANCE.getModLoader().isFabricLike()) {
+            ModelLoadingEvents.LOAD_MODEL.register(SpikeBlockModelHandler::onLoadModel);
+        }
+    }
 
     @Override
     public void onRegisterBlockEntityRenderers(BlockEntityRenderersContext context) {
@@ -17,6 +32,7 @@ public class SpikySpikesClient implements ClientModConstructor {
 
     @Override
     public void onRegisterItemModels(ItemModelsContext context) {
+        // TODO remove once the custom block model is supported for NeoForge
         context.registerSpecialModelRenderer(SpikySpikes.id("spike"), SpikeSpecialRenderer.Unbaked.MAP_CODEC);
     }
 }