+++ 1.21.4/Common/src/main/java/fuzs/spikyspikes/data/client/ModModelProvider.java
@@ -1,25 +1,19 @@
 package fuzs.spikyspikes.data.client;
 
 import fuzs.puzzleslib.api.client.data.v2.AbstractModelProvider;
-import fuzs.puzzleslib.api.client.data.v2.models.ModelTemplateHelper;
 import fuzs.puzzleslib.api.data.v2.core.DataProviderContext;
-import fuzs.spikyspikes.SpikySpikes;
 import fuzs.spikyspikes.client.renderer.special.SpikeSpecialRenderer;
 import fuzs.spikyspikes.init.ModRegistry;
 import fuzs.spikyspikes.world.level.block.SpikeBlock;
 import net.minecraft.client.data.models.BlockModelGenerators;
-import net.minecraft.client.data.models.model.ItemModelUtils;
-import net.minecraft.client.data.models.model.ModelTemplate;
-import net.minecraft.client.data.models.model.TextureMapping;
-import net.minecraft.client.data.models.model.TextureSlot;
+import net.minecraft.client.data.models.model.*;
 import net.minecraft.client.renderer.item.ItemModel;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 
 public class ModModelProvider extends AbstractModelProvider {
-    public static final ModelTemplate SPIKE_INVENTORY = ModelTemplateHelper.createItemModelTemplate(SpikySpikes.id(
-            "template_spike"), TextureSlot.PARTICLE);
+    public static final ModelTemplate BLOCK_MODEL_TEMPLATE = ModelTemplates.create("block", TextureSlot.PARTICLE);
 
     public ModModelProvider(DataProviderContext context) {
         super(context);
@@ -27,7 +21,7 @@ public class ModModelProvider extends AbstractModelProvider {
 
     @Override
     public void addBlockModels(BlockModelGenerators blockModelGenerators) {
-        this.createSpikeBlock(ModRegistry.WOODEN_SPIKE_BLOCK.value(), Blocks.OAK_PLANKS, blockModelGenerators);
+        this.createSpikeBlock(ModRegistry.WOODEN_SPIKE_BLOCK.value(), Blocks.STRIPPED_OAK_LOG, blockModelGenerators);
         this.createSpikeBlock(ModRegistry.STONE_SPIKE_BLOCK.value(), Blocks.SMOOTH_STONE, blockModelGenerators);
         this.createSpikeBlock(ModRegistry.IRON_SPIKE_BLOCK.value(), Blocks.IRON_BLOCK, blockModelGenerators);
         this.createSpikeBlock(ModRegistry.GOLDEN_SPIKE_BLOCK.value(), Blocks.GOLD_BLOCK, blockModelGenerators);
@@ -36,10 +30,12 @@ public class ModModelProvider extends AbstractModelProvider {
     }
 
     public final void createSpikeBlock(Block spikeBlock, Block particleBlock, BlockModelGenerators blockModelGenerators) {
-        blockModelGenerators.createParticleOnlyBlock(spikeBlock, particleBlock);
-        ResourceLocation resourceLocation = SPIKE_INVENTORY.create(spikeBlock.asItem(),
+        ResourceLocation resourceLocation = BLOCK_MODEL_TEMPLATE.create(spikeBlock,
                 TextureMapping.particle(particleBlock),
                 blockModelGenerators.modelOutput);
+        blockModelGenerators.blockStateOutput.accept(BlockModelGenerators.createSimpleBlock(spikeBlock,
+                resourceLocation).with(blockModelGenerators.createColumnWithFacing()));
+        // TODO remove this so the item model can use the block model once the custom model is supported for NeoForge
         ItemModel.Unbaked unbaked = ItemModelUtils.specialModel(resourceLocation,
                 new SpikeSpecialRenderer.Unbaked(((SpikeBlock) spikeBlock).getSpikeMaterial()));
         blockModelGenerators.itemModelOutput.accept(spikeBlock.asItem(), unbaked);