+++ 1.21.4/Common/src/main/java/fuzs/spikyspikes/init/ModRegistry.java
@@ -1,5 +1,6 @@
 package fuzs.spikyspikes.init;
 
+import fuzs.puzzleslib.api.data.v2.AbstractDatapackRegistriesProvider;
 import fuzs.puzzleslib.api.init.v3.registry.RegistryManager;
 import fuzs.puzzleslib.api.init.v3.registry.TransmuteRecipeHelper;
 import fuzs.spikyspikes.SpikySpikes;
@@ -8,7 +9,10 @@ import fuzs.spikyspikes.world.level.block.SpikeBlock;
 import fuzs.spikyspikes.world.level.block.SpikeMaterial;
 import fuzs.spikyspikes.world.level.block.entity.SpikeBlockEntity;
 import net.minecraft.core.Holder;
+import net.minecraft.core.RegistrySetBuilder;
 import net.minecraft.core.component.DataComponents;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.damagesource.DamageType;
 import net.minecraft.world.item.CreativeModeTab;
@@ -21,7 +25,12 @@ import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.material.MapColor;
 import net.minecraft.world.level.material.PushReaction;
 
+import java.util.Set;
+
 public class ModRegistry {
+    public static final RegistrySetBuilder REGISTRY_SET_BUILDER = new RegistrySetBuilder().add(Registries.DAMAGE_TYPE,
+            ModRegistry::bootstrapDamageTypes);
+
     static final RegistryManager REGISTRIES = RegistryManager.from(SpikySpikes.MOD_ID);
     public static final Holder.Reference<Block> WOODEN_SPIKE_BLOCK = REGISTRIES.whenOnFabricLike()
             .registerBlock("wooden_spike",
@@ -100,15 +109,19 @@ public class ModRegistry {
     public static final Holder.Reference<BlockEntityType<SpikeBlockEntity>> SPIKE_BLOCK_ENTITY_TYPE = REGISTRIES.registerBlockEntityType(
             "spike",
             SpikeBlockEntity::new,
-            WOODEN_SPIKE_BLOCK,
-            STONE_SPIKE_BLOCK,
-            IRON_SPIKE_BLOCK,
-            GOLDEN_SPIKE_BLOCK,
-            DIAMOND_SPIKE_BLOCK,
-            NETHERITE_SPIKE_BLOCK);
+            () -> Set.of(WOODEN_SPIKE_BLOCK.value(),
+                    STONE_SPIKE_BLOCK.value(),
+                    IRON_SPIKE_BLOCK.value(),
+                    GOLDEN_SPIKE_BLOCK.value(),
+                    DIAMOND_SPIKE_BLOCK.value(),
+                    NETHERITE_SPIKE_BLOCK.value()));
     public static final ResourceKey<DamageType> SPIKE_DAMAGE_TYPE = REGISTRIES.registerDamageType("spike");
 
     public static void bootstrap() {
         TransmuteRecipeHelper.registerTransmuteRecipeSerializers(REGISTRIES);
+    }
+
+    static void bootstrapDamageTypes(BootstrapContext<DamageType> context) {
+        AbstractDatapackRegistriesProvider.registerDamageType(context, SPIKE_DAMAGE_TYPE);
     }
 }