+++ 1.21.4/CHANGELOG.md
@@ -4,5 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.4.1-1.21.4] - 2025-03-16
+### Changed
+- Spike models now use baked models instead of entity model parts (Fabric only)
+### Fixed
+- Fix spikes spawning too many block particles when broken
+
 ## [v21.4.0-1.21.4] - 2025-02-10
 - Port to Minecraft 1.21.4